+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added tests for `terra-list`.
+
 ## 1.23.0 - (March 1, 2023)
 
 * Changed